Added delete list API functionality

+++ pages/api/lists/[listId].ts
@@ -71,23 +71,42 @@ const listApiRoutes = async (req, res) => {
 	}
 
 	if (req.method === 'DELETE') {
-		//Delete it from its respective list
-		const list = await List.findById(req.body.listId);
-		const newListItems = await list.items.filter(
-			(item) => item._id.toString() !== req.body.listItemId
-		);
-		list.items = newListItems;
-		await list.save();
+		if (req.body.action === 'delete-list') {
+			//Delete all of its items
+			const deletedListItems = await ListItem.deleteMany({
+				list: req.body.listId,
+			});
 
-		//Delete the list item itself
-		const deletedListIem = await ListItem.deleteOne({
-			_id: req.body.listItemId,
-		});
+			//Delete the list itself
+			const deletedList = await List.deleteOne({
+				_id: req.body.listId,
+			});
 
-		res.send({
-			status: 200,
-			data: deletedListIem,
-		});
+			res.send({
+				status: 200,
+				data: { deletedList, deletedListItems },
+			});
+		}
+
+		if (req.body.action === 'delete-list-item') {
+			//Delete it from its respective list
+			const list = await List.findById(req.body.listId);
+			const newListItems = await list.items.filter(
+				(item) => item._id.toString() !== req.body.listItemId
+			);
+			list.items = newListItems;
+			await list.save();
+
+			//Delete the list item itself
+			const deletedListIem = await ListItem.deleteOne({
+				_id: req.body.listItemId,
+			});
+
+			res.send({
+				status: 200,
+				data: deletedListIem,
+			});
+		}
 	}
 };
